assert useCart return type as tuple
