+++ packages/react-hooks/src/hooks/use-cart/use-cart.provider.tsx
@@ -1,7 +1,7 @@
 import React, { useReducer, useMemo, useContext, FC, ReactNode } from 'react';
 import { NacelleShopProduct, CartItem } from '@nacelle/types';
 
-import { CartState, CheckoutStatus } from './use-cart.types';
+import { CartState, CheckoutStatus, CartActions } from './use-cart.types';
 import cartReducer, {
   initialState,
   ADD_TO_CART,
@@ -13,17 +13,6 @@ import cartReducer, {
   SET_CHECKOUT_STATUS,
   CLEAR_CART
 } from './use-cart.reducer';
-
-export type CartActions = {
-  addToCart: (payload: NacelleShopProduct) => void;
-  updateItem: (payload: NacelleShopProduct | CartItem) => void;
-  removeFromCart: (payload: NacelleShopProduct | CartItem) => void;
-  incrementItem: (payload: NacelleShopProduct | CartItem) => void;
-  decrementItem: (payload: NacelleShopProduct | CartItem) => void;
-  setCheckoutStatus: (payload: CheckoutStatus) => void;
-  toggleCart: () => void;
-  clearCart: () => void;
-};
 
 export type CartProviderProps = {
   useLocalStorage: boolean;
